requirements.txt
prefect==0.15.11
prefect[kubernetes, github]==0.15.11
# azure-common==1.1.27
# azure-core==1.21.1
azure-identity==1.7.1
azure-keyvault-secrets==4.3.0
azure-storage-blob==12.9.0
azure-storage-file-datalake==12.5.0